df.append works with mixed categoricals
