+++ dask/dataframe/tests/test_dataframe.py
@@ -956,93 +956,6 @@ def test_unknown_divisions():
     assert_eq(d.a + d.b + 1, full.a + full.b + 1)
 
 
-def test_append():
-    df = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6],
-                       'b': [1, 2, 3, 4, 5, 6]})
-    df2 = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6],
-                        'b': [1, 2, 3, 4, 5, 6]},
-                       index=[6, 7, 8, 9, 10, 11])
-    df3 = pd.DataFrame({'b': [1, 2, 3, 4, 5, 6],
-                        'c': [1, 2, 3, 4, 5, 6]},
-                       index=[6, 7, 8, 9, 10, 11])
-
-    ddf = dd.from_pandas(df, 2)
-    ddf2 = dd.from_pandas(df2, 2)
-    ddf3 = dd.from_pandas(df3, 2)
-
-    s = pd.Series([7, 8], name=6, index=['a', 'b'])
-    assert_eq(ddf.append(s), df.append(s))
-
-    assert_eq(ddf.append(ddf2), df.append(df2))
-    assert_eq(ddf.a.append(ddf2.a), df.a.append(df2.a))
-    # different columns
-    assert_eq(ddf.append(ddf3), df.append(df3))
-    assert_eq(ddf.a.append(ddf3.b), df.a.append(df3.b))
-
-    # dask + pandas
-    assert_eq(ddf.append(df2), df.append(df2))
-    assert_eq(ddf.a.append(df2.a), df.a.append(df2.a))
-
-    assert_eq(ddf.append(df3), df.append(df3))
-    assert_eq(ddf.a.append(df3.b), df.a.append(df3.b))
-
-    df4 = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6],
-                        'b': [1, 2, 3, 4, 5, 6]},
-                       index=[4, 5, 6, 7, 8, 9])
-    ddf4 = dd.from_pandas(df4, 2)
-    msg = ("Unable to append two dataframes to each other with known "
-           "divisions if those divisions are not ordered. "
-           "The divisions/index of the second dataframe must be "
-           "greater than the divisions/index of the first dataframe.")
-    with tm.assertRaisesRegexp(ValueError, msg):
-        ddf.append(ddf4)
-
-
-def test_append2():
-    dsk = {('x', 0): pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6]}),
-           ('x', 1): pd.DataFrame({'a': [4, 5, 6], 'b': [3, 2, 1]}),
-           ('x', 2): pd.DataFrame({'a': [7, 8, 9], 'b': [0, 0, 0]})}
-    meta = make_meta({'a': 'i8', 'b': 'i8'})
-    ddf1 = dd.DataFrame(dsk, 'x', meta, [None, None])
-
-    dsk = {('y', 0): pd.DataFrame({'a': [10, 20, 30], 'b': [40, 50, 60]}),
-           ('y', 1): pd.DataFrame({'a': [40, 50, 60], 'b': [30, 20, 10]}),
-           ('y', 2): pd.DataFrame({'a': [70, 80, 90], 'b': [0, 0, 0]})}
-    ddf2 = dd.DataFrame(dsk, 'y', meta, [None, None])
-
-    dsk = {('y', 0): pd.DataFrame({'b': [10, 20, 30], 'c': [40, 50, 60]}),
-           ('y', 1): pd.DataFrame({'b': [40, 50, 60], 'c': [30, 20, 10]})}
-    meta = make_meta({'b': 'i8', 'c': 'i8'})
-    ddf3 = dd.DataFrame(dsk, 'y', meta, [None, None])
-
-    assert_eq(ddf1.append(ddf2), ddf1.compute().append(ddf2.compute()))
-    assert_eq(ddf2.append(ddf1), ddf2.compute().append(ddf1.compute()))
-    # Series + DataFrame
-    assert_eq(ddf1.a.append(ddf2), ddf1.a.compute().append(ddf2.compute()))
-    assert_eq(ddf2.a.append(ddf1), ddf2.a.compute().append(ddf1.compute()))
-
-    # different columns
-    assert_eq(ddf1.append(ddf3), ddf1.compute().append(ddf3.compute()))
-    assert_eq(ddf3.append(ddf1), ddf3.compute().append(ddf1.compute()))
-    # Series + DataFrame
-    assert_eq(ddf1.a.append(ddf3), ddf1.a.compute().append(ddf3.compute()))
-    assert_eq(ddf3.b.append(ddf1), ddf3.b.compute().append(ddf1.compute()))
-
-    # Dask + pandas
-    assert_eq(ddf1.append(ddf2.compute()), ddf1.compute().append(ddf2.compute()))
-    assert_eq(ddf2.append(ddf1.compute()), ddf2.compute().append(ddf1.compute()))
-    # Series + DataFrame
-    assert_eq(ddf1.a.append(ddf2.compute()), ddf1.a.compute().append(ddf2.compute()))
-    assert_eq(ddf2.a.append(ddf1.compute()), ddf2.a.compute().append(ddf1.compute()))
-
-    # different columns
-    assert_eq(ddf1.append(ddf3.compute()), ddf1.compute().append(ddf3.compute()))
-    assert_eq(ddf3.append(ddf1.compute()), ddf3.compute().append(ddf1.compute()))
-    # Series + DataFrame
-    assert_eq(ddf1.a.append(ddf3.compute()), ddf1.a.compute().append(ddf3.compute()))
-    assert_eq(ddf3.b.append(ddf1.compute()), ddf3.b.compute().append(ddf1.compute()))
-
-
 @pytest.mark.parametrize('join', ['inner', 'outer', 'left', 'right'])
 def test_align(join):
     df1a = pd.DataFrame({'A': np.random.randn(10),
